drop spree support
After trying to get the spree support working for the past two hours,
I’ve concluded that its just not worth it. I know there’s a lot of you
out there that want this. If you truly need to keep using Spree, please
fork the repo and make the necessary changes.